--- a/pyulog/core.py
+++ b/pyulog/core.py
@@ -579,16 +579,20 @@ class ULog(object):
         """
         sync_seq_found = False
         initial_file_position = self._file_handle.tell()
+        current_file_position = initial_file_position
 
         if last_n_bytes != -1:
             self._file_handle.seek(-last_n_bytes, 1)
+            current_file_position = current_file_position - last_n_bytes
 
         sync_start = self._file_handle.read(1)
+        current_file_position = current_file_position + 1
         try:
             while last_n_bytes == -1 or\
-             (self._file_handle.tell() < initial_file_position):
+             (current_file_position < initial_file_position):
                 if sync_start[0] == ULog.SYNC_BYTES[0]:
                     data = self._file_handle.read(7)
+                    current_file_position = current_file_position + 7
                     if data == ULog.SYNC_BYTES[1:]:
                         sync_seq_found = True
                         if self._debug:
@@ -599,7 +603,9 @@ class ULog(object):
                     else:
                         # seek back 7 bytes and look for sync start again
                         self._file_handle.seek(-7, 1)
+                        current_file_position = current_file_position - 7
                 sync_start = self._file_handle.read(1)
+                current_file_position = current_file_position + 1
         except IndexError:
             # Reached end of file
             if self._debug:
@@ -607,6 +613,7 @@ class ULog(object):
 
         if not sync_seq_found:
             self._file_handle.seek(initial_file_position, 0)
+            current_file_position = initial_file_position
 
             if last_n_bytes == -1:
                 self._has_sync = False