--- a/pyulog/core.py
+++ b/pyulog/core.py
@@ -587,25 +587,26 @@ class ULog(object):
             current_file_position = self._file_handle.seek(-last_n_bytes, 1)
             search_chunk_size = last_n_bytes
 
-        s = self._file_handle.read(search_chunk_size)
-        while len(s) > 0:
-            current_file_position += len(s)
-
-            k = s.find(ULog.SYNC_BYTES)
-            if k > 0:
-                print("Found sync sequence at [%i, %i]" %\
-                            (current_file_position - len(s) + k, current_file_position - len(s) + k + len(ULog.SYNC_BYTES)))
+        chunk = self._file_handle.read(search_chunk_size)
+        while len(chunk) > 0:
+            current_file_position += len(chunk)
+            chunk_index = chunk.find(ULog.SYNC_BYTES)
+            if chunk_index > 0:
+                if self._debug:
+                    print("Found sync at %i" % current_file_position - len(chunk) + chunk_index)
                 # seek to end of sync sequence and break
-                current_file_position = self._file_handle.seek(current_file_position - len(s) + k + len(ULog.SYNC_BYTES), 0)
+                current_file_position = self._file_handle.seek(current_file_position - len(chunk)\
+                         + chunk_index + len(ULog.SYNC_BYTES), 0)
                 sync_seq_found = True
                 break
 
             elif last_n_bytes != -1:
+                # we read the whole last_n_bytes and did not find sync
                 break
 
             else:
                 # read next chunk
-                s = self._file_handle.read(search_chunk_size)
+                chunk = self._file_handle.read(search_chunk_size)
 
         if not sync_seq_found:
             current_file_position = self._file_handle.seek(initial_file_position, 0)