--- a/pyulog/core.py
+++ b/pyulog/core.py
@@ -604,9 +604,9 @@ class ULog(object):
                 # we read the whole last_n_bytes and did not find sync
                 break
 
-            else:
-                # read next chunk
-                chunk = self._file_handle.read(search_chunk_size)
+            # seek back 7 bytes to handle boundary condition and read next chunk
+            current_file_position = self._file_handle.seek(-7, 1)
+            chunk = self._file_handle.read(search_chunk_size)
 
         if not sync_seq_found:
             current_file_position = self._file_handle.seek(initial_file_position, 0)