Speedup parsing by eliminating repeated calls to file_handle.tell().

Before:

```
In [12]: cProfile.run("ULog('/home/bharat/Downloads/final_obc2_bad2.ulg')")
         8063278 function calls (8063275 primitive calls) in 3.670 seconds

   Ordered by: standard name

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
     1694    0.001    0.000    0.001    0.000 <ipython-input-10-ae14ac574a72>:17(_parse_string)
        1    0.000    0.000    0.000    0.000 <ipython-input-10-ae14ac574a72>:195(has_data_appended)
       33    0.000    0.000    0.000    0.000 <ipython-input-10-ae14ac574a72>:226(__init__)
        2    0.000    0.000    0.000    0.000 <ipython-input-10-ae14ac574a72>:263(__init__)
   547023    0.134    0.000    0.175    0.000 <ipython-input-10-ae14ac574a72>:267(initialize)
      814    0.002    0.000    0.003    0.000 <ipython-input-10-ae14ac574a72>:273(__init__)
        1    0.000    0.000    0.000    0.000 <ipython-input-10-ae14ac574a72>:293(__init__)
      128    0.001    0.000    0.003    0.000 <ipython-input-10-ae14ac574a72>:309(__init__)
     1341    0.001    0.000    0.001    0.000 <ipython-input-10-ae14ac574a72>:318(_extract_type)
      612    0.001    0.000    0.003    0.000 <ipython-input-10-ae14ac574a72>:336(__init__)
      810    0.000    0.000    0.000    0.000 <ipython-input-10-ae14ac574a72>:378(__init__)
       33    0.001    0.000    0.002    0.000 <ipython-input-10-ae14ac574a72>:384(__init__)
       33    0.000    0.000    0.001    0.000 <ipython-input-10-ae14ac574a72>:408(_parse_format)
    36/33    0.001    0.000    0.001    0.000 <ipython-input-10-ae14ac574a72>:416(_parse_nested_type)
        1    0.000    0.000    0.000    0.000 <ipython-input-10-ae14ac574a72>:440(__init__)
   518833    0.418    0.000    0.511    0.000 <ipython-input-10-ae14ac574a72>:443(initialize)
       96    0.000    0.000    0.000    0.000 <ipython-input-10-ae14ac574a72>:463(_add_message_info_multiple)
        1    0.000    0.000    3.668    3.668 <ipython-input-10-ae14ac574a72>:473(_load_file)
        1    0.000    0.000    0.000    0.000 <ipython-input-10-ae14ac574a72>:498(_read_file_header)
        1    0.002    0.002    0.009    0.009 <ipython-input-10-ae14ac574a72>:511(_read_file_definitions)
    23348    0.535    0.000    1.601    0.000 <ipython-input-10-ae14ac574a72>:570(_find_sync)
        1    0.816    0.816    3.660    3.660 <ipython-input-10-ae14ac574a72>:622(_read_file_data)
    23348    0.006    0.000    0.006    0.000 <ipython-input-10-ae14ac574a72>:711(_check_packet_corruption)
     1694    0.001    0.000    0.002    0.000 <ipython-input-10-ae14ac574a72>:81(parse_string)
        1    0.000    0.000    3.669    3.669 <ipython-input-10-ae14ac574a72>:94(__init__)
        1    0.002    0.002    3.670    3.670 <string>:1(<module>)
        1    0.000    0.000    0.000    0.000 vt100_input.py:278(_input_parser_generator)
     2911    0.001    0.000    0.001    0.000 {built-in method _struct.unpack}
        1    0.000    0.000    3.670    3.670 {built-in method builtins.exec}
        1    0.000    0.000    0.000    0.000 {built-in method builtins.isinstance}
   547827    0.036    0.000    0.036    0.000 {built-in method builtins.len}
        1    0.000    0.000    0.000    0.000 {built-in method io.open}
       33    0.000    0.000    0.000    0.000 {built-in method numpy.core.multiarray.frombuffer}
     3598    0.000    0.000    0.000    0.000 {method 'append' of 'list' objects}
        1    0.000    0.000    0.000    0.000 {method 'close' of '_io.BufferedReader' objects}
        1    0.000    0.000    0.000    0.000 {method 'disable' of '_lsprof.Profiler' objects}
     1543    0.000    0.000    0.000    0.000 {method 'find' of 'str' objects}
       33    0.000    0.000    0.000    0.000 {method 'newbyteorder' of 'numpy.dtype' objects}
      103    0.000    0.000    0.000    0.000 {method 'pop' of 'list' objects}
       33    0.000    0.000    0.000    0.000 {method 'popitem' of 'dict' objects}
  2644694    0.301    0.000    0.301    0.000 {method 'read' of '_io.BufferedReader' objects}
    47993    0.011    0.000    0.011    0.000 {method 'seek' of '_io.BufferedReader' objects}
     2411    0.001    0.000    0.001    0.000 {method 'split' of 'str' objects}
      947    0.000    0.000    0.000    0.000 {method 'startswith' of 'str' objects}
  2106569    1.262    0.000    1.262    0.000 {method 'tell' of '_io.BufferedReader' objects}
  1584690    0.135    0.000    0.135    0.000 {method 'unpack' of 'Struct' objects}

```


After

```
cProfile.run("ULog('/home/bharat/Downloads/final_obc2_bad2.ulg')")
6526212 function calls (6526209 primitive calls) in 2.662 seconds

   Ordered by: standard name

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
     1694    0.002    0.000    0.002    0.000 <ipython-input-13-d5d38212270c>:17(_parse_string)
        1    0.000    0.000    0.000    0.000 <ipython-input-13-d5d38212270c>:195(has_data_appended)
       33    0.000    0.000    0.000    0.000 <ipython-input-13-d5d38212270c>:226(__init__)
        2    0.000    0.000    0.000    0.000 <ipython-input-13-d5d38212270c>:263(__init__)
   547023    0.129    0.000    0.170    0.000 <ipython-input-13-d5d38212270c>:267(initialize)
      814    0.006    0.000    0.012    0.000 <ipython-input-13-d5d38212270c>:273(__init__)
        1    0.000    0.000    0.000    0.000 <ipython-input-13-d5d38212270c>:293(__init__)
      128    0.003    0.000    0.010    0.000 <ipython-input-13-d5d38212270c>:309(__init__)
     1341    0.003    0.000    0.005    0.000 <ipython-input-13-d5d38212270c>:318(_extract_type)
      612    0.001    0.000    0.003    0.000 <ipython-input-13-d5d38212270c>:336(__init__)
      810    0.002    0.000    0.002    0.000 <ipython-input-13-d5d38212270c>:378(__init__)
       33    0.001    0.000    0.005    0.000 <ipython-input-13-d5d38212270c>:384(__init__)
       33    0.000    0.000    0.003    0.000 <ipython-input-13-d5d38212270c>:408(_parse_format)
    36/33    0.001    0.000    0.003    0.000 <ipython-input-13-d5d38212270c>:416(_parse_nested_type)
        1    0.000    0.000    0.000    0.000 <ipython-input-13-d5d38212270c>:440(__init__)
   518833    0.408    0.000    0.500    0.000 <ipython-input-13-d5d38212270c>:443(initialize)
       96    0.000    0.000    0.000    0.000 <ipython-input-13-d5d38212270c>:463(_add_message_info_multiple)
        1    0.000    0.000    2.661    2.661 <ipython-input-13-d5d38212270c>:473(_load_file)
        1    0.000    0.000    0.000    0.000 <ipython-input-13-d5d38212270c>:498(_read_file_header)
        1    0.007    0.007    0.033    0.033 <ipython-input-13-d5d38212270c>:511(_read_file_definitions)
    23348    0.382    0.000    0.549    0.000 <ipython-input-13-d5d38212270c>:570(_find_sync)
        1    0.851    0.851    2.628    2.628 <ipython-input-13-d5d38212270c>:629(_read_file_data)
    23348    0.006    0.000    0.006    0.000 <ipython-input-13-d5d38212270c>:718(_check_packet_corruption)
     1694    0.002    0.000    0.004    0.000 <ipython-input-13-d5d38212270c>:81(parse_string)
        1    0.000    0.000    2.661    2.661 <ipython-input-13-d5d38212270c>:94(__init__)
        1    0.002    0.002    2.662    2.662 <string>:1(<module>)
        1    0.000    0.000    0.000    0.000 vt100_input.py:278(_input_parser_generator)
     2911    0.002    0.000    0.002    0.000 {built-in method _struct.unpack}
        1    0.000    0.000    2.662    2.662 {built-in method builtins.exec}
        1    0.000    0.000    0.000    0.000 {built-in method builtins.isinstance}
   547827    0.036    0.000    0.036    0.000 {built-in method builtins.len}
        1    0.000    0.000    0.000    0.000 {built-in method io.open}
       33    0.000    0.000    0.000    0.000 {built-in method numpy.core.multiarray.frombuffer}
     3598    0.001    0.000    0.001    0.000 {method 'append' of 'list' objects}
        1    0.000    0.000    0.000    0.000 {method 'close' of '_io.BufferedReader' objects}
        1    0.000    0.000    0.000    0.000 {method 'disable' of '_lsprof.Profiler' objects}
     1543    0.001    0.000    0.001    0.000 {method 'find' of 'str' objects}
       33    0.000    0.000    0.000    0.000 {method 'newbyteorder' of 'numpy.dtype' objects}
      103    0.000    0.000    0.000    0.000 {method 'pop' of 'list' objects}
       33    0.000    0.000    0.000    0.000 {method 'popitem' of 'dict' objects}
  2644694    0.294    0.000    0.294    0.000 {method 'read' of '_io.BufferedReader' objects}
    47993    0.012    0.000    0.012    0.000 {method 'seek' of '_io.BufferedReader' objects}
     2411    0.003    0.000    0.003    0.000 {method 'split' of 'str' objects}
      947    0.001    0.000    0.001    0.000 {method 'startswith' of 'str' objects}
   569503    0.372    0.000    0.372    0.000 {method 'tell' of '_io.BufferedReader' objects}
  1584690    0.133    0.000    0.133    0.000 {method 'unpack' of 'Struct' objects}

```

Note that so many calls to find sync only happen when you do have unknown message type in your ULog file.