@devbharat 2019-06-13
@bkueng Also note that this block also considerably slows down parsing

```
                if self._file_handle.tell() > read_until:
                    if self._debug:
                        print('read until offset=%i done, current pos=%i' %
                              (read_until, self._file_handle.tell()))
                    break
```

With it: 5033340 function calls (5033337 primitive calls) in 2.049 seconds
Without: 4487186 function calls (4487183 primitive calls) in 1.565 seconds