@bkueng 2019-06-14
Cool. For me it's still slow, and I see this (on a log w/o sync messages):

```
   Ordered by: cumulative time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
    203/1    0.002    0.000    7.213    7.213 {built-in method builtins.exec}
        1    0.000    0.000    7.213    7.213 ulog_info:3(<module>)
        1    0.000    0.000    6.794    6.794 info.py:68(main)
        1    0.000    0.000    6.792    6.792 core.py:94(__init__)
        1    0.000    0.000    6.792    6.792 core.py:473(_load_file)
        1    0.013    0.013    6.786    6.786 core.py:631(_read_file_data)
        1    5.113    5.113    6.756    6.756 core.py:572(_find_sync)
 15683145    1.641    0.000    1.641    0.000 {method 'read' of '_io.BufferedReader' objects}
        9    0.002    0.000    0.521    0.058 __init__.py:1(<module>)
    233/3    0.003    0.000    0.416    0.139 <frozen importlib._bootstrap>:978(_find_and_load)
    233/3    0.002    0.000    0.416    0.139 <frozen importlib._bootstrap>:948(_find_and_load_unlocked)
    220/4    0.002    0.000    0.416    0.104 <frozen importlib._bootstrap>:663(_load_unlocked)
    178/4    0.001    0.000    0.416    0.104 <frozen importlib._bootstrap_external>:722(exec_module)
    305/4    0.000    0.000    0.415    0.104 <frozen importlib._bootstrap>:211(_call_with_frames_removed)
   359/18    0.001    0.000    0.263    0.015 {built-in method builtins.__import__}
        1    0.000    0.000    0.221    0.221 __init__.py:16(<module>)
        1    0.000    0.000    0.197    0.197 __init__.py:485(load_entry_point)
        1    0.000    0.000    0.196    0.196 __init__.py:2723(load_entry_point)
        1    0.000    0.000    0.195    0.195 __init__.py:2333(load)
        1    0.000    0.000    0.193    0.193 __init__.py:2348(resolve)
  888/167    0.001    0.000    0.189    0.001 <frozen importlib._bootstrap>:1009(_handle_fromlist)
```

So `_find_sync` is slow, and `BufferedReader.read` uses a portion of that, but it's more that the loop is just slow (due to it being python and iterating 1 byte at a time). I think we could speed it up by reading and searching in a junk of data. This is worth a test as well: https://stackoverflow.com/a/3219275.