@devbharat 2019-06-14
> on a log w/o sync messages

True, I didn't test that case!