@devbharat 2019-06-14
The last commit fixed this case (no sync bytes in file) for me:

Before: 22304291 function calls (22304288 primitive calls) in 6.312 seconds
After: 5698780 function calls (5698777 primitive calls) in 1.974 seconds

Probably needs another thorough review and testing.