@devbharat 2019-06-17
> There's a corner case: if the sync bytes are split over 2 consecutive chunks, they are not found.

Yes i am aware(mentioned on the old sync PR slow discussion), we could do a preventive seek back by 7 bytes after every failed sync search in every chunk, but that'll slow it down a little. Perhaps it is worth it.