@bkueng 2019-06-17
We can also add the last bytes of the previous chunk to the next one to avoid seeking.