@bkueng 2019-06-24
Rebased, cleaned up the commits and merged. Thanks for fixing and testing!